requirements.txt
python-telegram-bot==11.1.0
langdetect==1.0.9